+++ src/main/java/com/coniverse/dangjang/domain/auth/service/AuthTokenGenerator.java
@@ -1,11 +1,11 @@
-package com.coniverse.dangjang.domain.auth.service.authToken;
+package com.coniverse.dangjang.domain.auth.service;
 
 import java.util.Date;
 
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 
-import com.coniverse.dangjang.domain.auth.dto.AuthToken.AuthToken;
+import com.coniverse.dangjang.domain.auth.dto.AuthToken;
 
 import lombok.RequiredArgsConstructor;
 
@@ -13,12 +13,12 @@ import lombok.RequiredArgsConstructor;
  * JWT 생성 및 추출
  *
  * @author EVE
- * @since 1.0
+ * @since 1.0.0
  */
 
-@Service
+@Component
 @RequiredArgsConstructor
-public class AuthTokensGenerator {
+public class AuthTokenGenerator {
 	private static final String BEARER_TYPE = "Bearer";
 	@Value("${jwt.access.exp}")
 	private long accessTokenExpireTime;
@@ -28,18 +28,17 @@ public class AuthTokensGenerator {
 	private final JwtTokenProvider jwtTokenProvider;
 
 	/**
-	 * JWT 생성
+	 * auth token을 생성한다
 	 *
-	 * @param memberId 유저 ID
-	 * @return AuthTokens JWT TOKEN 반환
-	 * @since 1.0
+	 * @param subject 유저 id
+	 * @return AuthToken jwt token
+	 * @since 1.0.0
 	 */
-	public AuthToken generate(String memberId) {
-		long now = (new Date()).getTime();
+	public AuthToken generate(String subject) {
+		long now = new Date().getTime();
 		Date accessTokenExpiredAt = new Date(now + accessTokenExpireTime);
 		Date refreshTokenExpiredAt = new Date(now + refreshTokenExpireTime);
 
-		String subject = memberId;
 		String accessToken = jwtTokenProvider.generate(subject, accessTokenExpiredAt);
 		String refreshToken = jwtTokenProvider.generate(subject, refreshTokenExpiredAt);
 
@@ -47,11 +46,11 @@ public class AuthTokensGenerator {
 	}
 
 	/**
-	 * JWT에서 유저 ID 추출
+	 * access token에서 유저 id를 추출한다
 	 *
 	 * @param accessToken JWT 토큰
-	 * @return Long 유저ID 반환
-	 * @since 1.0
+	 * @return user id
+	 * @since 1.0.0
 	 */
 	public Long extractMemberId(String accessToken) {
 		return Long.valueOf(jwtTokenProvider.extractSubject(accessToken));
